move window-title to it's own folder (as it's a shared component)
