+++ src/electron/views/common/window-title.tsx
